flask_app/requirements.txt

flask
python-dotenv
requests
vosk
#google-generative-ai
pyttsx3
gunicorn
